# services/runtime/internal/adapter/http/openapi.yaml
openapi: 3.0.3
info:
  title: Backify Runtime API
  description: |
    Data-plane service — Auth end-user, CRUD, ownership-based permission.
    Mọi request (trừ /health, /internal/*, /swagger/*) cần header **X-Project-Id**.
  version: 0.1.0
servers:
  - url: http://localhost:8081
    description: Local

tags:
  - name: Health
  - name: Auth
  - name: CRUD
  - name: Internal

paths:
  /health:
    get:
      tags: [Health]
      summary: Health check
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  /v1/auth/signup:
    post:
      tags: [Auth]
      summary: End-user signup
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                fullName:
                  type: string
                phone:
                  type: string
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"

  /v1/auth/signin:
    post:
      tags: [Auth]
      summary: End-user signin
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          $ref: "#/components/responses/Error"

  /v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh tokens
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "401":
          $ref: "#/components/responses/Error"

  /v1/auth/me:
    get:
      tags: [Auth]
      summary: Current user
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Error"

  /v1/{entity}:
    post:
      tags: [CRUD]
      summary: Create record
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/Entity"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
            example:
              title: My order
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "403":
          $ref: "#/components/responses/Error"
    get:
      tags: [CRUD]
      summary: List records (ownership filtered)
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/Entity"
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListResult"

  /v1/{entity}/{id}:
    get:
      tags: [CRUD]
      summary: Get one record
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/Entity"
        - $ref: "#/components/parameters/RecordId"
      responses:
        "200":
          description: OK
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    patch:
      tags: [CRUD]
      summary: Update record (owner only)
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/Entity"
        - $ref: "#/components/parameters/RecordId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: OK
        "403":
          $ref: "#/components/responses/Error"
    delete:
      tags: [CRUD]
      summary: Delete record (owner only)
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/Entity"
        - $ref: "#/components/parameters/RecordId"
      responses:
        "204":
          description: Deleted
        "403":
          $ref: "#/components/responses/Error"

  /internal/bootstrap/{projectID}:
    post:
      tags: [Internal]
      summary: Bootstrap project schema
      security:
        - InternalKey: []
      parameters:
        - name: projectID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK

  /internal/config/{projectID}:
    get:
      tags: [Internal]
      summary: Debug project config
      security:
        - InternalKey: []
      parameters:
        - name: projectID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK

  /internal/projects/{projectID}/users/{userID}/role:
    post:
      tags: [Internal]
      summary: Promote user role (admin/user)
      security:
        - InternalKey: []
      parameters:
        - name: projectID
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: userID
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Internal-Key
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role:
                  type: string
                  enum: [admin, user]
      responses:
        "204":
          description: Updated

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    InternalKey:
      type: apiKey
      in: header
      name: X-Internal-Key

  parameters:
    ProjectId:
      name: X-Project-Id
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: Project UUID
    Entity:
      name: entity
      in: path
      required: true
      schema:
        type: string
      description: Entity name (e.g. Order, User)
      example: Order
    RecordId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid

  schemas:
    TokenPair:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        fullName:
          type: string
        phone:
          type: string
        role:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        tokens:
          $ref: "#/components/schemas/TokenPair"
    ListResult:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            additionalProperties: true
        total:
          type: integer
    Error:
      type: object
      properties:
        error:
          type: string
        code:
          type: string

  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
            